Struct file is created
